Revert "Revert "Merge remote-tracking branch 'origin/main' into core_cuda""
This reverts commit 5e01f462e184ad189dffe02483aacc984963e066.

+++ src/cpp/dfm.h
@@ -30,10 +30,9 @@ using namespace std;
 template <typename T>
 py::array_t<double> dfm_direct(py::array_t<T, py::array::c_style> img_seq,
                                vector<unsigned int> lags,
-                               size_t nx,
-                               size_t ny);
+                               unsigned long long nx,
+                               unsigned long long ny);
 
-                               
 
 /*! \brief Compute ISF in fft mode using Wiener-Khinchin theorem
     \param img_seq      numpy array containing the image sequence
@@ -46,10 +45,11 @@ py::array_t<double> dfm_direct(py::array_t<T, py::array::c_style> img_seq,
 template <typename T>
 py::array_t<double> dfm_fft(py::array_t<T, py::array::c_style> img_seq,
                             vector<unsigned int> lags,
-                            size_t nx,
-                            size_t ny,
-                            size_t nt,
-                            size_t chunk_size);
+                            unsigned long long nx,
+                            unsigned long long ny,
+                            unsigned long long nt,
+                            unsigned long long chunk_size);
+
 
 /*! \brief Export dfm functions to python
     \param m    Module
